+++ source/package.js
@@ -595,11 +595,6 @@ export async function updatePackageData(state) {
 	)
 
 	// engines
-	if (answers.website) {
-		packageData.engines.node = `>=${answers.desiredNodeVersion}`
-	} else {
-		packageData.engines.node = `>=${answers.nodeVersionSupportedMinimum}`
-	}
 	delete packageData.engines.docpad
 
 	// license